create levelLogUtils
